msm: clock-8x60: Vote for mmfpb_a_clk at 64MHz
Previously, no subsystem voted for the multimedia fast
peripheral bus clock.  As a result, it remained at what the
bootloaders set it to (64MHz). Assert an active-only vote
from Apps so that the clock rate can be dropped in power-
collapse.

The MMFPB will still be at 64MHz whenever an application
processor is running, but will now drop to 27MHz when
they are power-collapsed.

(cherry picked from commit 5b0d9be)

Change-Id: I26a1153230b5ffb5b73fb462f6e5f84ea3632f81
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>