msm: rpc: While powering down flush all the rpcrouter workqueues
When the device is powered down, RPC Router closes all the transports
and releases all the associated data structures. This sometimes lead to
works queued in the RPC router workqueues to access stale data and hence
cause kernel panic. So when the device is powered down, abort the works
queued in the workqueues and wait until the workqueues are flushed and
then free the associated data structures.

CRs-Fixed: 312078
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@codeaurora.org>
(cherry picked from commit c74ab97)

Change-Id: Ia342637156df0cc3f934bc7778c2ef94dda735c3
Signed-off-by: Shruthi Krishna <skrish@codeaurora.org>
modified by faux123