msm_fb: display: Flush RGB pipe when solid fill mode is used
Fullscreen applications can request the driver to configure the MDP
back ground pipe to use solid fill mode so that the background pipe
will not be fetched by MDP and it saves the bus bandwidth hence the
total system power. This requires the driver to explicitly issue a
register flush on the background pipe when the configuration changes
from regular fetch to solid fill mode during overlay play.

CRs-Fixed: 270713
Signed-off-by: Rajesh Sastrula <vrajesh@codeaurora.org>